+++ package.json
@@ -11,6 +11,7 @@
     "dependencies": {
         "@prisma/client": "^3.14.0",
         "@types/react-js-pagination": "^3.0.4",
+        "crypto": "^1.0.1",
         "form-data": "^4.0.0",
         "gray-matter": "^4.0.3",
         "iron-session": "^6.1.3",
